+++ src/tbrugz/sqldump/datadump/DumpSyntax.java
@@ -36,7 +36,7 @@ public abstract class DumpSyntax {
 	
 	public abstract void dumpHeader(Writer fos) throws Exception;
 
-	public abstract void dumpRow(ResultSet rs, int count, Writer fos) throws Exception;
+	public abstract void dumpRow(ResultSet rs, long count, Writer fos) throws Exception;
 
 	public abstract void dumpFooter(Writer fos) throws Exception;
