+++ src/tbrugz/sqldump/datadump/HTMLDataDump.java
@@ -50,7 +50,7 @@ public class HTMLDataDump extends DumpSyntax {
 	}
 
 	@Override
-	public void dumpRow(ResultSet rs, int count, Writer fos) throws Exception {
+	public void dumpRow(ResultSet rs, long count, Writer fos) throws Exception {
 		StringBuffer sb = new StringBuffer();
 		sb.append("\t"+"<tr>");
 		List vals = SQLUtils.getRowObjectListFromRS(rs, lsColTypes, numCol);
